Move validator module to identity_credential
