+++ identity_credential/src/validator/errors.rs
@@ -4,6 +4,8 @@
 use std::collections::BTreeMap;
 use std::fmt::Display;
 
+use itertools;
+
 #[derive(Debug, thiserror::Error, strum::IntoStaticStr)]
 #[non_exhaustive]
 /// An error associated with validating credentials and presentations.
@@ -14,26 +16,32 @@ pub enum ValidationError {
   /// Indicates that the issuance date of the credential is not considered valid.
   #[error("issuance date is in the future or later than required")]
   IssuanceDate,
-  /// Indicates that the credential's (resp. presentation's) signature could not be verified using the issuer's (resp.
-  /// holder's) DID Document.
+  /// Indicates that the credential's (resp. presentation's) signature could not be verified using
+  /// the issuer's (resp. holder's) DID Document.
   #[error("could not verify the {signer_ctx}'s signature")]
   #[non_exhaustive]
   Signature {
+    /// Signature verification error.
     source: Box<dyn std::error::Error + Send + Sync + 'static>,
+    /// Specifies whether the error was from the DID Document of a credential issuer
+    /// or the presentation holder.
     signer_ctx: SignerContext,
   },
 
-  /// Indicates that the credential's (resp. presentation's) issuer's (resp. holder's) URL could not be parsed as a
-  /// valid DID.
+  /// Indicates that the credential's (resp. presentation's) issuer's (resp. holder's) URL could
+  /// not be parsed as a valid DID.
   #[error("{signer_ctx} URL is not a valid DID")]
   #[non_exhaustive]
   SignerUrl {
+    /// DID parsing error.
     source: Box<dyn std::error::Error + Send + Sync + 'static>,
+    /// Specifies whether the error relates to the DID of a credential issuer
+    /// or the presentation holder.
     signer_ctx: SignerContext,
   },
 
-  /// Indicates an attempt to verify a signature of a credential (resp. presentation) using a DID Document not matching
-  /// the issuer's (resp. holder's) id.
+  /// Indicates an attempt to verify a signature of a credential (resp. presentation) using a
+  /// DID Document not matching the issuer's (resp. holder's) id.
   #[error("the {0}'s id does not match the provided DID Document(s)")]
   #[non_exhaustive]
   DocumentMismatch(SignerContext),
@@ -41,11 +49,11 @@ pub enum ValidationError {
   /// Indicates that the structure of the [Credential](identity_credential::credential::Credential) is not semantically
   /// correct.
   #[error("the credential's structure is not semantically correct")]
-  CredentialStructure(#[source] identity_credential::Error),
+  CredentialStructure(#[source] crate::Error),
   /// Indicates that the structure of the [Presentation](identity_credential::presentation::Presentation) is not
   /// semantically correct.
   #[error("the presentation's structure is not semantically correct")]
-  PresentationStructure(#[source] identity_credential::Error),
+  PresentationStructure(#[source] crate::Error),
   /// Indicates that the relationship between the presentation holder and one of the credential subjects is not valid.
   #[error("expected holder = subject of the credential")]
   #[non_exhaustive]
@@ -55,7 +63,7 @@ pub enum ValidationError {
   MissingPresentationHolder,
   /// Indicates that the credential's status is invalid.
   #[error("invalid credential status")]
-  InvalidStatus(#[source] identity_credential::Error),
+  InvalidStatus(#[source] crate::Error),
   /// Indicates that the the credential's service is invalid.
   #[error("invalid service")]
   InvalidService(#[source] identity_did::Error),
@@ -64,10 +72,13 @@ pub enum ValidationError {
   Revoked,
 }
 
+/// Specifies whether an error is related to a credential issuer or the presentation holder.
 #[derive(Debug)]
 #[non_exhaustive]
 pub enum SignerContext {
+  /// Credential issuer.
   Issuer,
+  /// Presentation holder.
   Holder,
 }
 
@@ -81,9 +92,10 @@ impl Display for SignerContext {
   }
 }
 
+/// Errors caused by a failure to validate a credential.
 #[derive(Debug)]
-/// An error caused by a failure to validate a Credential.  
 pub struct CompoundCredentialValidationError {
+  /// List of credential validation errors.
   pub validation_errors: Vec<ValidationError>,
 }
 
@@ -104,7 +116,10 @@ impl std::error::Error for CompoundCredentialValidationError {}
 #[derive(Debug)]
 /// An error caused by a failure to validate a Presentation.
 pub struct CompoundPresentationValidationError {
+  /// Errors that occurred during validation of individual credentials, mapped by index of their
+  /// order in the presentation.
   pub credential_errors: BTreeMap<usize, CompoundCredentialValidationError>,
+  /// Errors that occurred during validation of the presentation.
   pub presentation_validation_errors: Vec<ValidationError>,
 }
