+++ identity_did/src/revocation/bitmap.rs
@@ -147,7 +147,7 @@ impl<D: DID + Sized, T> TryFrom<&Service<D, T>> for RevocationBitmap {
   fn try_from(service: &Service<D, T>) -> Result<Self> {
     if service.type_() != Self::TYPE {
       return Err(Error::InvalidService(
-        "invalid service - expected a `RevocationBitmap2022`",
+        "invalid type - expected `RevocationBitmap2022`",
       ));
     }
