+++ identity_iota_client/src/lib.rs
@@ -20,7 +20,6 @@ pub use self::error::Error;
 pub use self::error::Result;
 
 pub mod chain;
-pub mod credential;
 pub mod document;
 pub mod tangle;
